refactor: Comment 컴포넌트를 분리했다
