+++ src/pages/ChannelPage.js
@@ -67,7 +67,7 @@ const LinkButton = styled.button`
 
 function ChannelPage() {
   const { user, isLogin } = useValueContext();
-  const userId = user && user._id; // 전역스토어에서 가져옴
+  const userId = user && user._id;
   const { favorites } = useActionContext();
   const navigate = useNavigate();
   const [start, setStart] = useState(0);
